% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleaning_text_function}
\alias{cleaning_text_function}
\title{Clean data scrapped from reddit}
\usage{
cleaning_text_function(x, stopwords)
}
\arguments{
\item{x}{A \code{dataframe} (dataframe) we want to clean}

\item{stopwords}{A \code{list} (list) used to specify the stopwords we want to take off the dataframe}
}
\value{
A \code{dataframe} with the scraped data cleaned
}
\description{
Clean data scrapped from reddit
}
\section{Functions}{
\itemize{
\item \code{cleaning_text_function}: returns a dataframe cleaned
}}

