% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_plot.R
\name{comment_plot}
\alias{comment_plot}
\title{Comment_plot}
\usage{
comment_plot(df)
}
\arguments{
\item{analyse}{the comment on reddit}

\item{adapted}{dataframe}
}
\value{
plot and analysis of the comment
}
\description{
Comment_plot
}
