% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_user_comments}
\alias{get_user_comments}
\title{Get comments from a user}
\usage{
get_user_comments(user = "", page_threshold = 2, wait_time = 4)
}
\arguments{
\item{user}{A \code{char} (character) used to specify what the username}

\item{wait_time}{A \code{num} (numeric) used to specify the waiting time between scrappings.}

\item{page_treshold}{A \code{num} (numeric) used to specify the maxiumum number of pages to scrap}
}
\value{
A \code{dataframe} with the scraping done
}
\description{
Get comments from a user
}
\section{Functions}{
\itemize{
\item \code{get_user_comments}: returns a dataframe scrapped using username.
}}

