% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reddit.R
\name{map_reddit}
\alias{map_reddit}
\title{Get a map with countries that appear in comments}
\usage{
map_reddit(df)
}
\arguments{
\item{word}{a word to research reddit for}
}
\value{
map
}
\description{
Get a map with countries that appear in comments
}
