% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montage.R
\name{montage}
\alias{montage}
\title{Get a mosaique from image in the subreddit}
\usage{
montage(df)
}
\arguments{
\item{specify}{a subreddit}
}
\value{
pane of images that show what are images within the subreddit
}
\description{
Get a mosaique from image in the subreddit
}
