% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sentiments_reddit.R
\name{plot_sentiments_reddit}
\alias{plot_sentiments_reddit}
\title{Plot sentiments from a researched word}
\usage{
plot_sentiments_reddit(df)
}
\arguments{
\item{word}{used to specify the word to research reddit for}

\item{stopwords}{used to specify a list of words that will be used as stopwords}
}
\value{
sentiments plot
}
\description{
Plot sentiments from a researched word
}
