% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud_reddit.R
\name{plot_wordcloud_reddit}
\alias{plot_wordcloud_reddit}
\title{Get wordcloud}
\usage{
plot_wordcloud_reddit(df)
}
\arguments{
\item{word}{used to specify the word to research reddit for}

\item{stopwordsused}{to specify a list of words that will be used as stopwords}
}
\value{
with the wordcloud
}
\description{
Get wordcloud
}
