% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_wordcloudReddit}
\alias{plot_wordcloudReddit}
\title{Get wordcloud}
\usage{
plot_wordcloudReddit(word, stopwords)
}
\arguments{
\item{word}{used to specify the word to research reddit for}

\item{stopwordsused}{to specify a list of words that will be used as stopwords}
}
\value{
with the wordcloud
}
\description{
Get wordcloud
}
\section{Functions}{
\itemize{
\item \code{plot_wordcloudReddit}: returns a wordcloud of a typed word
}}

