% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reddit_content}
\alias{reddit_content}
\title{Retrieve content from given URL}
\usage{
reddit_content(URL, wait_time = 2)
}
\arguments{
\item{URL}{A \code{URL} (URL) used to specify what URL we want to extract the data from}

\item{wait_time}{A \code{num} (numeric) used to specify the time between scrappings}
}
\value{
A \code{dataframe} with the scraping done
}
\description{
Retrieve content from given URL
}
\section{Functions}{
\itemize{
\item \code{reddit_content}: returns a dataframe scrapped using a url.
}}

