% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_function.R
\name{scrapping_function}
\alias{scrapping_function}
\title{Scrap data from reddit to get a table}
\usage{
scrapping_function(search_terms, subreddit, sort_by, time_frame)
}
\arguments{
\item{search_terms}{The word we look for}

\item{subreddit}{the subreddit we search in}

\item{sort_by}{the sort by new, top...}

\item{time_frame}{the time frame taken into account}
}
\value{
the scraped data cleaned
}
\description{
Scrap data from reddit to get a table
}
