% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiments_per_hour.R
\name{sentiments_per_hour}
\alias{sentiments_per_hour}
\title{Get sentiments per hour of the day}
\usage{
sentiments_per_hour(df)
}
\arguments{
\item{b}{a database extracted from reddit_content()}
}
\value{
plot
}
\description{
Get sentiments per hour of the day
}
