% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_text_function.R
\name{cleaning_text_function}
\alias{cleaning_text_function}
\title{Clean data scrapped from reddit}
\usage{
cleaning_text_function(x, stopwords = stopwords_vec)
}
\arguments{
\item{x}{we want to clean}

\item{stopwords}{used to specify the stopwords we want to take off the dataframe}
}
\value{
the scraped data cleaned
}
\description{
Clean data scrapped from reddit
}
