% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_comments.R
\name{get_user_comments}
\alias{get_user_comments}
\title{Get comments from a user}
\usage{
get_user_comments(user = "", page_threshold = 5, wait_time = 2)
}
\arguments{
\item{user}{used to specify what the username}

\item{wait_time}{used to specify the waiting time between scrappings.}

\item{page_treshold}{used to specify the maxiumum number of pages to scrap}
}
\value{
A dataframe with the scraping done
}
\description{
Get comments from a user
}
