% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_sentimentsReddit}
\alias{plot_sentimentsReddit}
\title{Plot sentiments from a researched word}
\usage{
plot_sentimentsReddit(word, stopwords)
}
\arguments{
\item{word}{used to specify the word to research reddit for}

\item{stopwords}{used to specify a list of words that will be used as stopwords}
}
\value{
sentiments
}
\description{
Plot sentiments from a researched word
}
\section{Functions}{
\itemize{
\item \code{plot_sentimentsReddit}: returns a sentiment analysis of a typed word .
}}

