% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_content.R
\name{reddit_content}
\alias{reddit_content}
\title{Retrieve content from given URL}
\usage{
reddit_content(URL, wait_time = 1)
}
\arguments{
\item{URL}{used to specify what URL we want to extract the data from}

\item{wait_time}{used to specify the time between scrappings}
}
\value{
A \code{dataframe} with the scraping done
}
\description{
Retrieve content from given URL
}
