% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_urls_mod.R
\name{reddit_urls_mod}
\alias{reddit_urls_mod}
\title{Scrapping urls}
\usage{
reddit_urls_mod(
  search_terms = "",
  subreddit = "",
  sort_by = "",
  time_frame = ""
)
}
\arguments{
\item{search_terms}{A \code{char} (character) used to specify what the user is looking for}

\item{subreddit}{A \code{char} (character) used to specify what subreddit we want to scrape from. Default is NA}

\item{sort_by}{A \code{char} (character) used to specify if we scrape by new, by relevance etc. Relevance is the default}

\item{time_frame}{A \code{char} (character) used to specify if we look only at last week, last month, or all}
}
\value{
A \code{dataframe} with the scraping done
}
\description{
Scrapping urls
}
